Enable the possibility to read either self-trigger or full-streaming
These changes affect to the WaveformSet.from_ROOT_file() class method, which is an alternative initializer. The changes introduced may be counterproductive, from the run-time point of view, depending on how mixed the self-trigger samples and the full-streaming samples are in the input bulk-data tree. Future commits may introduce a new input parameter to this method to let the user decide whether it is better to, either perform a block-by-block reading or, just read the whole arrays and later discard the data that is not desired.